+++ TestMap/Services/Database/Repositories/AnalysisSolutionRepository.cs
@@ -35,8 +35,8 @@ public class AnalysisSolutionRepository
             using var reader = await checkCmd.ExecuteReaderAsync();
             if (await reader.ReadAsync())
             {
-                int id = reader.GetInt16(0);
-                int projectId = reader.GetInt16(1);
+                int id = reader.GetInt32(0);
+                int projectId = reader.GetInt32(1);
                 var guid = reader.GetString(2);
 
                 solution.Id = id;
