+++ TestMap/Services/Database/Repositories/ClassRepository.cs
@@ -33,7 +33,7 @@ public class ClassRepository
         await using var reader = await checkCmd.ExecuteReaderAsync();
         if (await reader.ReadAsync())
         {
-            int id = reader.GetInt16(0);
+            int id = reader.GetInt32(0);
             var guid = reader.GetString(1);
 
             cla.Id = id;
