+++ TestMap/Services/Database/Repositories/CoverageReportRepository.cs
@@ -33,7 +33,7 @@ public class CoverageReportRepository
         using var reader = await checkCmd.ExecuteReaderAsync();
         if (await reader.ReadAsync())
         {
-            int id = reader.GetInt16(0);
+            int id = reader.GetInt32(0);
 
             return id;
         }
@@ -83,7 +83,7 @@ public class CoverageReportRepository
         using var reader = await checkCmd.ExecuteReaderAsync();
         if (await reader.ReadAsync())
         {
-            int id = reader.GetInt16(0);
+            int id = reader.GetInt32(0);
 
             if (id > 0)
             {
