+++ TestMap/Services/Database/Repositories/LizardFunctionCodeMetricsRepository.cs
@@ -59,7 +59,7 @@ public class LizardFunctionCodeMetricsRepository
         using var reader = await checkCmd.ExecuteReaderAsync();
         if (await reader.ReadAsync())
         {
-            int id = reader.GetInt16(0);
+            int id = reader.GetInt32(0);
 
             if (id > 0)
             {
