+++ TestMap/Services/Database/Repositories/MethodRepository.cs
@@ -34,7 +34,7 @@ public class MethodRepository
         using var reader = await checkCmd.ExecuteReaderAsync();
         if (await reader.ReadAsync())
         {
-            int id = reader.GetInt16(0);
+            int id = reader.GetInt32(0);
             var guid = reader.GetString(1);
 
             method.Id = id;
