+++ TestMap/Services/Database/Repositories/MutationReportRepository.cs
@@ -61,7 +61,7 @@ public class MutationReportRepository
         using var reader = await checkCmd.ExecuteReaderAsync();
         if (await reader.ReadAsync())
         {
-            int id = reader.GetInt16(0);
+            int id = reader.GetInt32(0);
 
             if (id > 0)
             {
